pyproject.toml
# pyproject.toml

[build-system]
# requires = ["setuptools>=61.0.0", "wheel"]
# build-backend = "setuptools.build_meta"
requires = ["flit_core>=3.4"]
build-backend = "flit_core.buildapi"

[project]
name = "micromind"
dynamic=["version"]
description = "MicroMind"
readme = "README.md"
authors = [{ name = "Francesco Paissan & others", email = "francescopaissan@gmail.com" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["feed", "reader", "tutorial"]
dependencies = [
    "torch",
    "torchvision",
    "torchaudio",
    "torchinfo",
    "huggingface_hub",
    "accelerate==0.23.0",
    "onnx",
    "loguru"
]
requires-python = ">=3.8"

[project.optional-dependencies]
dev = ["black", "bumpver", "flake8", "isort", "pip-tools", "pytest", "pre-commit"]
conversion = [
    "onnx",
    "onnx_tf==1.10.0",
    "onnx-simplifier==0.4.13",
    "onnxruntime==1.13.1",
    "openvino-dev==2022.3.0",
    "openvino2tensorflow==1.34.0",
    "tensorflow_datasets==4.8.1",
    "tensorflow==2.11.0",
    "tensorflow_probability==0.19.0"
]

[project.urls]
Homepage = "https://github.com/fpaissan/micromind"

[project.scripts]
phinet = "__main__:main"

[tool.isort]
profile = "black"

[tool.setuptools]
py-modules = []

[tool.bumpver]
current_version = "0.2.1"

version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "bump version {old_version} -> {new_version}"
commit = true
tag = true
push = false


[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
]
"micromind/__init__.py" = [
    "{version}",
]
